Consolidate walker helper into xml-pull bootstrap
* Reduce file requirements by moving the helper class into the only
client that uses it
